refactored some for better readibility, and error resolving
