+++ src/components/Themed.tsx
@@ -6,7 +6,7 @@ import React from 'react';
 import { Text as DefaultText, View as DefaultView } from 'react-native';
 
 import Colors from '../constants/Colors';
-import useColorScheme from '../hooks/useColorScheme';
+import useColorScheme from '../utils/hooks/useColorScheme';
 
 export function useThemeColor(
   props: { light?: string; dark?: string },
