+++ src/components/texts/TextInputDetail.tsx
